+++ packages/router/src/private_export.ts
@@ -11,4 +11,4 @@ export {RestoredState as ɵRestoredState} from './navigation_transition';
 export {loadChildren as ɵloadChildren} from './router_config_loader';
 export {ROUTER_PROVIDERS as ɵROUTER_PROVIDERS} from './router_module';
 export {afterNextNavigation as ɵafterNextNavigation} from './utils/navigations';
-export {withActivatedRouteInjectors as ɵwithActivatedRouteInjectors} from './provide_router';
+export {withRouterResources as ɵwithActivatedRouteInjectors} from './provide_router';
